requirements.txt
requests
pyaudio
wave
python-dotenv
google-cloud-speech
google-cloud-texttospeech
pydub
